+++ core/src/main/java/nl/nn/adapterframework/jdbc/dbms/IDbmsSupport.java
@@ -1,5 +1,5 @@
 /*
-   Copyright 2013, 2015, 2018 Nationale-Nederlanden
+   Copyright 2013, 2015, 2018, 2019 Nationale-Nederlanden
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -91,6 +91,8 @@ public interface IDbmsSupport {
 	String provideTrailingFirstRowsHint(int rowCount);
 
 	String getSchema(Connection conn) throws JdbcException;
+	
+	String convertQuery(Connection conn, String query, String dbmsFrom) throws SQLException, JdbcException;
 
 	boolean isTablePresent(Connection conn, String tableName) throws JdbcException;
 	boolean isTableColumnPresent(Connection conn, String schemaName, String tableName, String columnName) throws JdbcException;